% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_eusilc.R
\name{merge_eusilc}
\alias{merge_eusilc}
\title{Merge eu-silc personal level cross-sectional raw .csv datasets, personal_register and personal_data}
\usage{
merge_eusilc(path.personal.register, path.personal.data,
  path.household.register, path.household.data, level, output.path,
  output.filename = "default", format, subset.vars.per.reg = "all",
  subset.vars.per.data = "all", subset.vars.hh.reg = "all",
  subset.vars.hh.data = "all", subset.countries = "all",
  return.object = TRUE)
}
\arguments{
\item{path.personal.register}{A string. Specify the path to original personal register .csv file.}

\item{path.personal.data}{A string. Specify the path to original personal data .csv file.}

\item{path.household.register}{A string. Specify the path to original household register .csv file.}

\item{path.household.data}{A string. Specify the path to original household data .csv file.}

\item{level}{A string. Specify the whether to merge \code{"personal"}, \code{"household"} or \code{"both"} level datas. Both stands for merging all the four datas at personal level}

\item{output.path}{A string. Specify the path where to save the merged file. value=\code{"not_save"} does not save the merged file}

\item{output.filename}{A string. Specify filename for the merged file. A file extension according to format will be added.}

\item{format}{A string. Specify the output format for merged data. values=\code{"csv","RData","SPSS","SAS" or "Stata"}}

\item{subset.countries}{A string. Specify subset of countries. In \code{c("FI","SE")} format.  \code{"all"} includes all the countries.}

\item{return.object}{A logical. Whether function should return merged data.frame into environment.}

\item{subset.vars.per.reg.}{A string. Specify subset of variables from personal register file. \code{"all"} includes all the variables.}

\item{subset.vars.per.data.}{A string. Specify subset of variables from personal data file. \code{"all"} includes all the variables.}

\item{subset.vars.hh.reg.}{A string. Specify subset of variables from household register file. \code{"all"} includes all the variables.}

\item{subset.vars.hh.data.}{A string. Specify subset of variables from household data file. \code{"all"} includes all the variables.}
}
\value{
data.frame
}
\description{
Merge eu-silc personal level cross-sectional raw .csv datasets, personal_register and personal_data
}
\examples{
# 
}
\author{
Markus Kainu <markuskainu(at)gmail.com>
}

